refactor(commands): extract abstract commands
extract abstract commands to their respective subdirectories as `Base.ts`
