+++ src/commands/Reddit/meme.ts
@@ -1,9 +1,7 @@
-import {
-	AstraeaRedditCommand,
-	AstraeaCommandOptions
-} from '../../lib/Structures/Command'
-import { PieceContext } from '@sapphire/framework'
 import { ApplyOptions } from '@sapphire/decorators'
+import { PieceContext } from '@sapphire/framework'
+import { AstraeaCommandOptions } from '../../lib/Structures/Command'
+import { AstraeaRedditCommand } from './Base'
 
 @ApplyOptions<AstraeaCommandOptions>({
 	name: 'meme',
